Refactor timestamp types from uint32_t to size_t for consistency across input and recording systems

+++ src/InputSystem.cpp
@@ -315,7 +315,7 @@ std::vector<std::string> InputSystem::GetAvailableKeys() const {
     return keys;
 }
 
-void InputSystem::Apply(unsigned char *keyboardState, uint32_t currentTick) {
+void InputSystem::Apply(unsigned char *keyboardState, size_t currentTick) {
     if (!keyboardState || !m_Enabled) {
         return;
     }
